CommonAPI-D-Bus 3.1.8

+++ src/pugixml/pugixml.cpp
@@ -7583,7 +7583,7 @@ PUGI__NS_BEGIN
 			
 				if (expr->rettype() == xpath_type_number)
 				{
-					if (expr->eval_number(c, stack) == i)
+					if (expr->eval_number(c, stack) == (double)i)
 						*last++ = *it;
 				}
 				else if (expr->eval_boolean(c, stack))
@@ -8445,7 +8445,7 @@ PUGI__NS_BEGIN
 				double first = round_nearest(_right->eval_number(c, stack));
 				
 				if (is_nan(first)) return xpath_string(); // NaN
-				else if (first >= s_length + 1) return xpath_string();
+				else if (first >= (double)(s_length + 1)) return xpath_string();
 				
 				size_t pos = first < 1 ? 1 : static_cast<size_t>(first);
 				assert(1 <= pos && pos <= s_length + 1);
@@ -8468,12 +8468,12 @@ PUGI__NS_BEGIN
 				double last = first + round_nearest(_right->_next->eval_number(c, stack));
 				
 				if (is_nan(first) || is_nan(last)) return xpath_string();
-				else if (first >= s_length + 1) return xpath_string();
+				else if (first >= (double)(s_length + 1)) return xpath_string();
 				else if (first >= last) return xpath_string();
 				else if (last < 1) return xpath_string();
 				
 				size_t pos = first < 1 ? 1 : static_cast<size_t>(first);
-				size_t end = last >= s_length + 1 ? s_length + 1 : static_cast<size_t>(last);
+				size_t end = last >= (double)(s_length + 1) ? s_length + 1 : static_cast<size_t>(last);
 
 				assert(1 <= pos && pos <= end && end <= s_length + 1);
 				const char_t* rbegin = s.c_str() + (pos - 1);
